### 2021-02-06 #3016 @richford
At-mentioning co-authors here: @arokem, @mnarayan, @nrs02004, @jyeatman

Here is a non-exhaustive list of potential reviewers from the JOSS reviewers google sheet (coauthors, please feel free to add more in this thread):
- rougier
- sealhuang
- nirum
- effigies
- vsoch
- janfreyberg
- ionlights
- djsaunde
- JonathanReardon
- EtienneCmb
- ryEllison
- anibalsolon
- syoh
- clinssen
- emdupre
- saran-nns
