### 2021-02-08 #3016 @richford
Hi @galessiorob, thanks! I'm a bit confused about the missing DOI. The single missing DOI is also listed in the "OK DOIs" list.
