### 2021-02-09 #3016 @galessiorob
@richford you're right 🤔

@danielskatz do you think [this](https://github.com/openjournals/joss-reviews/issues/3016#issuecomment-775539130) can be a whedon bug?
