### 2021-02-09 #3016 @danielskatz
I wouldn't call this a bug.

In the .bib file, I see
> @inproceedings{mosci2010solving,
  title        = {Solving structured sparsity regularization with proximal methods},
  author       = {Mosci, Sofia and Rosasco, Lorenzo and Santoro, Matteo and Verri, Alessandro and Villa, Silvia},
  booktitle    = {Joint European conference on machine learning and knowledge discovery in databases},
  pages        = {418--433},
  year         = {2010},
  organization = {Springer}
}
>
> @inproceedings{mosci2010solving,
  author    = {Mosci, Sofia and Rosasco, Lorenzo and Santoro, Matteo and Verri, Alessandro and Villa, Silvia},
  editor    = {Balc{\'a}zar, Jos{\'e} Luis and Bonchi, Francesco and Gionis, Aristides and Sebag, Mich{\`e}le},
  title     = {{Solving Structured Sparsity Regularization with Proximal Methods}},
  booktitle = {{Machine Learning and Knowledge Discovery in Databases}},
  year      = {2010},
  publisher = {Springer Berlin Heidelberg},
  address   = {Berlin, Heidelberg},
  pages     = {418--433},
  isbn      = {978-3-642-15883-4},
  doi       = {10.1007/978-3-642-15883-4_27}
}

Both of these entries have the same title.  whedon saw the title of the first one, which doesn't have a DOI listed, looked it up, and found the DOI for the second one, which has the same title and appears to be the same work, or perhaps a different version of the same work (same title, same authors, same year).  And they have the same bibtex key.  Which one is correct and is intended to be cited in the paper?  The other should probably be deleted from the .bib file.
