### 2021-02-09 #3024 @whedon
[REVIEW]: Groupyr: Sparse Group Lasso in Python
**Submitting author:** @richford (<a href="http://orcid.org/0000-0001-9276-9084">Adam Richie-Halford</a>)
**Repository:** <a href="https://github.com/richford/groupyr" target ="_blank">https://github.com/richford/groupyr</a>
**Version:** v0.2.2
