### 2021-02-15 #3024 @janfreyberg
@richford can you please add citations & references for the group_lasso and Lightning packages in the paper? Thanks!
