### 2021-02-17 #3024 @rougier
I've just read the paper and installed the software. I've a few questions concerning the submission:

- There are five authors in the paper, could you tell us their respective role in the software?
- I did not  find the way to run the unittests. Would it be possible to have a simple `groupyrr.test()` that run tests?
- When you write "we found that our implementation .... was faster", maybe you can give some precision (2x 10x 100x ?) and possibly how to test it.
- For the references, I think some DOI are missing.
- For the reference on group-lasso, maybe you can retrieve the swid from software heritage such as to have a long term id on the resource.
