### 2021-02-17 #3024 @richford
Hi @rougier, thanks for your review. I'll address all but one of you points now and come back for the remaining one (about the performance comparison with group-lasso) in a few days:

> - There are five authors in the paper, could you tell us their respective role in the software?

Sure thing, in [this PR](https://github.com/richford/groupyr/pull/42), I added an author contribution statement to the paper, which categorizes author contribution according to the [CRediT taxonomy](http://dx.doi.org/10.1087/20150211).

> - I did not find the way to run the unittests. Would it be possible to have a simple groupyrr.test() that run tests?

Good question! Provided you have installed a development version of `groupyr` using
```
python -m pip install -e .[dev]
```
you can run unit tests and check for code style using
```
make lint
make test
```
We mention this in [this section of the contributing guidelines](https://github.com/richford/groupyr/blob/main/.github/CONTRIBUTING.md#making-a-change) (specifically , but your comment was very helpful because it made me realize that the provided `Makefile` is not well documented. To resolve this, I added a `help` target to the `Makefile` in [this recent PR](https://github.com/richford/groupyr/pull/43). Now, a user can simple type `make` or `make help` to see the available options and discover the `make test` command.

> - For the references, I think some DOI are missing.

You are correct. Thanks! I have added missing DOIs in [the same PR that added the author contribution statement](https://github.com/richford/groupyr/pull/42).

> - For the reference on group-lasso, maybe you can retrieve the swid from software heritage such as to have a long term id on the resource.

Thanks for the tip. I updated the `group-lasso` citation to use the biblatex `softwareversion` entry and then put the swhid in the version number. That change is also in [PR 42](https://github.com/richford/groupyr/pull/42)
