### 2021-02-17 #3024 @rougier
Thanks for your answers. For the speed comparison, I'm not asking to a full-proof comparison but just some number to give the reader an idea of the speedup (x1.1 is technically faster but 5x might be worth to explore further your software). If you have a script, it's all the better bit not mandatory at all.

By the way, I forgot to ask to list dependencies in the documenation. Not mandatory but it might help user.

Apart from that, I'm satisfied with the submission.
