### 2021-02-18 #3024 @richford
Hi @rougier, I addressed the speed comparison in [PR 44](https://github.com/richford/groupyr/pull/44), where I changed the comparison sentence to
> However, we found that our implementation in *groupyr*, which relies on the
*copt* optimization library, was between two and ten times faster
for the problem sizes that we encounter in our research (see the
repository's examples directory for a performance comparison).

And added an example script to the examples directory that generates a chart showing the speedup for various problem sizes.
