### 2021-02-18 #3024 @richford
@rougier, I also added a list of dependencies to the installation documentation in [PR 45](https://github.com/richford/groupyr/pull/45). Hope that resolves all of your concerns. Thanks!
