### 2021-02-21 #3024 @richford
Thanks to all of the reviewers: @rougier, @JonathanReardon, and @janfreyberg.

@galessiorob, it looks like all of the reviewers have completed their reviews. Anticipating the next steps, I have cut a new version of *groupyr* and created a new archive DOI. Here's the new information:

new version: 0.2.1
new DOI: [10.5281/zenodo.4554452](https://zenodo.org/record/4554452)

If you need any other information, please don't hesitate to reach out.
