### 2021-02-23 #3024 @richford
Sure thing. Sorry about that. I scraped the DOI [from the url](https://dl.acm.org/doi/10.5555/1953048.2078195) for the JMLR paper, but I guess that doesn't work. I see that later papers from the same sklearn authors reference their earlier paper and do not include a DOI at all. So I elected to simply remove the DOI for the sklearn reference. I just pushed that change to the `main` branch.

@galessiorob, do you want me to cut a new version because of this change? The underlying software did not change, only the `paper.bib` file that is in the repo.
