### 2021-02-24 #3024 @galessiorob
>@galessiorob, do you want me to cut a new version because of this change? The underlying software did not change, only the paper.bib file that is in the repo.

@richford thanks for the explanation, makes sense. Yes, let's get a new version 👍
