### 2021-02-24 #3024 @richford
@galessiorob, Thanks! The new version is `v0.2.2` and the new DOI is [10.5281/zenodo.4559599](https://zenodo.org/record/4559599)
