### 2021-02-24 #3024 @kthyng
Congrats on your new publication @richford!! Many thanks to editor @galessiorob and to reviewers @janfreyberg, @JonathanReardon, and @rougier — we couldn't do this without your hard work, time, and expertise!

(I'll close this issue once the doi resolves)
